requirements.txt
apscheduler>=3.11.0
bcrypt>=4.0.0
fastapi>=0.118.0
plotly>=6.3.1
sqlmodel>=0.0.25
streamlit-mic-recorder>=0.0.8
streamlit>=1.50.0
groq>=0.32.0
gtts>=2.5.4
chromadb>=1.2.2
pandas>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pytz>=2024.1
